# build cache hash using each_with_object